requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
supabase
python-multipart
